PR comments from Wolf and Bharath
