+++ Android/app/src/androidTest/java/PageObjects/TestAppPageObjects/NativeAppPageObject.kt
@@ -26,18 +26,17 @@
  */
 package PageObjects
 
-import android.os.Build
-import android.support.test.InstrumentationRegistry
-import android.support.test.uiautomator.UiDevice
+import android.support.test.uiautomator.UiSelector
+import org.junit.Assert
 
 /**
  * Created by bpage on 2/24/18.
  */
-open class BasePageObject {
-    val device: UiDevice = UiDevice.getInstance(InstrumentationRegistry.getInstrumentation())
-    // FIXME Update when min verison increaes past API 23
-    val isOldDevice: Boolean = (Build.VERSION.SDK_INT <= Build.VERSION_CODES.M)
-    // FIXME Update this when we stop using ARM Emulators
-    val isArm = Build.SUPPORTED_ABIS.first().contains("armeabi")
-    var timeout:Long = if (isArm) 30000 else 5000
+class NativeAppPageObject(private val app: TestApplication) : BasePageObject() {
+
+    fun assertAppLoads() {
+        val titleBar = device.findObject(UiSelector().className("android.widget.TextView").index(0))
+        titleBar.waitForExists(timeout * 10)
+        Assert.assertEquals("App did not successfully testLogin.", app.name, titleBar.text)
+    }
 }
