+++ Android/app/src/androidTest/java/PageObjects/TestAppPageObjects/ReactNativeAppPageObject.kt
@@ -26,18 +26,31 @@
  */
 package PageObjects
 
-import android.os.Build
-import android.support.test.InstrumentationRegistry
-import android.support.test.uiautomator.UiDevice
+import android.support.test.uiautomator.UiSelector
+import android.util.Log
+import org.junit.Assert
 
 /**
- * Created by bpage on 2/24/18.
+ * Created by bpage on 3/6/18.
  */
-open class BasePageObject {
-    val device: UiDevice = UiDevice.getInstance(InstrumentationRegistry.getInstrumentation())
-    // FIXME Update when min verison increaes past API 23
-    val isOldDevice: Boolean = (Build.VERSION.SDK_INT <= Build.VERSION_CODES.M)
-    // FIXME Update this when we stop using ARM Emulators
-    val isArm = Build.SUPPORTED_ABIS.first().contains("armeabi")
-    var timeout:Long = if (isArm) 30000 else 5000
+
+class ReactNativeAppPageObject : BasePageObject() {
+
+    init {
+        timeout *= 3
+    }
+
+    fun assertAppLoads() {
+        var alertWindow = device.findObject(UiSelector().resourceId("android:id/alertTitle"))
+        if (alertWindow.exists()) {
+            Log.i("uia", "React Native requesting overlay permission.")
+            // Tap Continue Button
+            device.findObject(UiSelector().resourceId("android:id/button1")).click()
+            Thread.sleep(timeout)
+        }
+
+        val title = device.findObject(UiSelector().className("android.widget.TextView").index(0))
+        title.waitForExists(timeout * 5)
+        Assert.assertEquals("App did not successfully login.", "Mobile SDK Sample App", title.text)
+    }
 }
