+++ Android/app/src/androidTest/java/com/salesforce/mobilesdk/mobilesdkuitest/Login/LoginTests.kt
@@ -1,3 +1,29 @@
+/*
+ * Copyright (c) 2017-present, salesforce.com, inc.
+ * All rights reserved.
+ * Redistribution and use of this software in source and binary forms, with or
+ * without modification, are permitted provided that the following conditions
+ * are met:
+ * - Redistributions of source code must retain the above copyright notice, this
+ * list of conditions and the following disclaimer.
+ * - Redistributions in binary form must reproduce the above copyright notice,
+ * this list of conditions and the following disclaimer in the documentation
+ * and/or other materials provided with the distribution.
+ * - Neither the name of salesforce.com, inc. nor the names of its contributors
+ * may be used to endorse or promote products derived from this software without
+ * specific prior written permission of salesforce.com, inc.
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+ * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+ * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
+ * ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE
+ * LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
+ * CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
+ * SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
+ * INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
+ * CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
+ * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
+ * POSSIBILITY OF SUCH DAMAGE.
+ */
 package com.salesforce.mobilesdk.mobilesdkuitest.Login
 
 import PageObjects.*
@@ -42,16 +68,10 @@ class LoginTests {
             AppType.NATIVE_JAVA, AppType.NATIVE_KOTLIN ->
                 NativeAppPageObject(app).assertAppLoads()
             AppType.HYBRID_LOCAL -> {
-                val title = device.findObject(UiSelector().className("android.view.View").descriptionContains("Users"))
-                title.waitForExists(timeout)
-                Assert.assertEquals(failedLoginMessage, "Users", title.contentDescription)
+                HybridLocalAppPageObject(app).assertAppLoads()
             }
             AppType.HYBRID_REMOTE -> {
-                Thread.sleep(timeout)
-                val title = device.findObject(UiSelector().className("android.view.View").descriptionContains("Salesforce Mobile SDK Test"))
-                title.waitForExists(timeout)
-                Thread.sleep(timeout / 2)
-                Assert.assertEquals(failedLoginMessage, "Salesforce Mobile SDK Test", title.contentDescription)
+                HybridRemoteAppPageObject(app).assertAppLoads()
             }
             AppType.REACT_NATIVE -> {
                 ReactNativeAppPageObject().assertAppLoads()
