+++ install.sh
@@ -5,7 +5,7 @@ if [[ "$OSTYPE" == "linux-gnu" ]]; then
     sudo apt-get install libqt5widgets5
 fi
 
-if [[ -v "CIRCLECI" ]]; then
+if [ -n "$CIRCLECI" ]; then
     sudo npm install -g cordova@7.0.0
     sudo npm install -g sfdx-cli
 else
